Refactored ExercisesTable.tsx - separated logic to a new file - useExercisesTableLogic.tsx, in ExercisesTable.tsx left only the view, moved function validateUrl do separate file and folder "helpers", changed files with logic extension to .ts
