+++ src/helpers/validateUrl.ts
@@ -1,0 +1,8 @@
+export const validateURL = (url: string) => {
+    try {
+        new URL(url);
+        return true;
+    } catch (error) {
+        return false;
+    }
+};
